client, response: Improve comparison formatting
