+++ main.go
@@ -123,9 +123,6 @@ func run(ctx context.Context, options Options) error {
 		timeout = d
 	}
 
-	ctx, done := context.WithCancel(ctx)
-	defer done()
-
 	g, ctx := errgroup.WithContext(ctx)
 
 	// Launch clients
@@ -145,7 +142,7 @@ func run(ctx context.Context, options Options) error {
 
 	if len(options.Verbose) > 0 {
 		r.MismatchedResponse = func(resps []Response) {
-			logger.Info().Msgf("mismatched responses: %v", resps)
+			logger.Info().Msgf("mismatched responses: %s", Responses(resps).String())
 		}
 	}
